refactor: 내부 설정 규격에 따라 Jpa 패키지 정정
