+++ uqs-domain/src/main/kotlin/com/dongholab/uqs/domain/member/service/MemberService.kt
@@ -1,13 +1,12 @@
 package com.dongholab.uqs.domain.member.service
 
 import com.dongholab.uqs.domain.member.Member
-import com.dongholab.uqs.domain.member.repository.MemberRepository
+import com.dongholab.uqs.domain.member.repository.mysql.MemberJpaRepository
 import org.springframework.stereotype.Service
 
-
 @Service
-class MemberService(private val repository: MemberRepository) {
+class MemberService(private val memberJpaRepository: MemberJpaRepository) {
     fun findByUserid(userId: String): Member? {
-        return repository.findByUserid(userId)
+        return memberJpaRepository.findByUserid(userId)
     }
 }
